automation/requirements.txt
playwright>=1.40.0
feedparser>=6.0.0
openai>=1.12.0
python-dotenv>=1.0.0
